Input buffers redesign
Input buffer architecture redesign:
- Now a buffer can track its physical and virtual size. This allows to check bounds when accessing virtual data, which was not possible before.
- Now a buffer can track if it is stateless or not. Image which uses stateless buffers is thread safe (when calling `const` methods).
- Added necessary classes and new tests.

Also added rebase tests and relocation entry tests and added new load config directory field support.

+++ tests/tests/pe_bliss2/directories/relocation_entry_tests.cpp
@@ -40,7 +40,7 @@ TEST(RelocationEntryTests, AffectedSizeTests)
 	EXPECT_EQ(entry.get_affected_size_in_bytes(), 0u);
 
 	entry.set_type(relocation_type::highadj);
-	EXPECT_EQ(entry.get_affected_size_in_bytes(), sizeof(std::uint32_t));
+	EXPECT_EQ(entry.get_affected_size_in_bytes(), sizeof(std::uint16_t));
 
 	entry.set_type(relocation_type::highlow);
 	EXPECT_EQ(entry.get_affected_size_in_bytes(), sizeof(std::uint32_t));
